lib: sbi: Fix incorrect SCExcl check
The previous condition checked whether there are no permissions for
the given cell at all.
However, for exclusive access with regard to given permissions, it
is sufficient to check whether there is an overlap of permissions.
For example, when requesting exclusive write access to a cell,
other security divisions can still have read access.